No whitespace in `navigationItemRelated` Graphql (#168)
Graphql type names can't contain whitespace. However, in Strapi the display name is a beautified, and human-friendly version of the content type which may contain spaces. This just removes them.

For example, a content type called "Policy Page" will become "PolicyPage" which is the Graphql type that strapi will generate